Move modules to make room for "random" AD scheme
