+++ exe/wallet/http-bridge/Main.hs
@@ -72,7 +72,7 @@ import Cardano.Wallet.Network
     ( NetworkLayer, defaultRetryPolicy, waitForConnection )
 import Cardano.Wallet.Primitive.AddressDerivation
     ( KeyToAddress )
-import Cardano.Wallet.Primitive.AddressDiscovery
+import Cardano.Wallet.Primitive.AddressDiscovery.Sequential
     ( SeqState )
 import Cardano.Wallet.Version
     ( showVersion, version )
