+++ lib/core/src/Cardano/Wallet/DB/Sqlite/TH.hs
@@ -38,8 +38,8 @@ import GHC.Generics
 import Numeric.Natural
     ( Natural )
 
-import qualified Cardano.Wallet.Primitive.AddressDerivation as W
-import qualified Cardano.Wallet.Primitive.AddressDiscovery as W
+import qualified Cardano.Wallet.Primitive.AddressDerivation.Sequential as W
+import qualified Cardano.Wallet.Primitive.AddressDiscovery.Sequential as W
 import qualified Cardano.Wallet.Primitive.Types as W
 import qualified Data.ByteString.Char8 as B8
