+++ lib/core/src/Cardano/Wallet/Primitive/AddressDerivation.hs
@@ -20,15 +20,14 @@
 --
 -- Primitives for performing address derivation for some given schemes. This is
 -- where most of the crypto happens in the wallet and, it is quite important to
--- ensure that the following implementation matches with other wallet softwares
--- (like Yoroi/Icarus or the cardano-cli)
+-- ensure that the implementations match with other Cardano wallets
+-- (like cardano-sl, Yoroi/Icarus, or cardano-cli)
 
 module Cardano.Wallet.Primitive.AddressDerivation
     (
     -- * Polymorphic / General Purpose Types
     -- $use
-      Key
-    , getKey
+      Key (..)
     , Depth (..)
     , Index (..)
     , DerivationType (..)
@@ -51,14 +50,6 @@ module Cardano.Wallet.Primitive.AddressDerivation
     , checkPassphrase
     , changePassphrase
 
-    -- * Sequential Derivation
-    , ChangeChain(..)
-    , generateKeyFromSeed
-    , unsafeGenerateKeyFromSeed
-    , deriveAccountPrivateKey
-    , deriveAddressPrivateKey
-    , deriveAddressPublicKey
-
     -- * Storing and retrieving keys
     , serializeXPrv
     , deserializeXPrv
@@ -69,19 +60,7 @@ module Cardano.Wallet.Primitive.AddressDerivation
 import Prelude
 
 import Cardano.Crypto.Wallet
-    ( DerivationScheme (..)
-    , XPrv
-    , XPub
-    , deriveXPrv
-    , deriveXPub
-    , generateNew
-    , toXPub
-    , unXPrv
-    , unXPub
-    , xPrvChangePass
-    , xprv
-    , xpub
-    )
+    ( XPrv, XPub, toXPub, unXPrv, unXPub, xPrvChangePass, xprv, xpub )
 import Cardano.Wallet.Primitive.Mnemonic
     ( CheckSumBits
     , ConsistentEntropy
@@ -95,7 +74,7 @@ import Cardano.Wallet.Primitive.Mnemonic
     , mnemonicToEntropy
     )
 import Cardano.Wallet.Primitive.Types
-    ( Address (..), Hash (..), invariant )
+    ( Address (..), Hash (..) )
 import Control.Arrow
     ( left )
 import Control.DeepSeq
@@ -118,22 +97,12 @@ import Data.ByteString
     ( ByteString )
 import Data.List
     ( intercalate )
-import Data.Maybe
-    ( fromMaybe )
 import Data.Proxy
     ( Proxy (..) )
 import Data.Text
     ( Text )
 import Data.Text.Class
-    ( CaseStyle (..)
-    , FromText (..)
-    , TextDecodingError (..)
-    , ToText (..)
-    , fromTextToBoundedEnum
-    , toTextFromBoundedEnum
-    )
-import Data.Typeable
-    ( Typeable )
+    ( FromText (..), TextDecodingError (..), ToText (..) )
 import Data.Word
     ( Word32 )
 import Fmt
@@ -440,167 +409,9 @@ changePassphrase (Passphrase oldPwd) (Passphrase newPwd) (Key prv) =
 instance MonadRandom ((->) (Passphrase "salt")) where
     getRandomBytes _ (Passphrase salt) = BA.convert salt
 
-
-
 {-------------------------------------------------------------------------------
-                            Sequential Derivation
+                          Backend-dependent functions
 -------------------------------------------------------------------------------}
-
--- | Marker for the change chain. In practice, change of a transaction goes onto
--- the addresses generated on the internal chain, whereas the external chain is
--- used for addresses that are part of the 'advertised' targets of a transaction
-data ChangeChain
-    = ExternalChain
-    | InternalChain
-    deriving (Generic, Typeable, Show, Eq, Ord, Bounded)
-
-instance NFData ChangeChain
-
--- Not deriving 'Enum' because this could have a dramatic impact if we were
--- to assign the wrong index to the corresponding constructor (by swapping
--- around the constructor above for instance).
-instance Enum ChangeChain where
-    toEnum = \case
-        0 -> ExternalChain
-        1 -> InternalChain
-        _ -> error "ChangeChain.toEnum: bad argument"
-    fromEnum = \case
-        ExternalChain -> 0
-        InternalChain -> 1
-
-instance ToText ChangeChain where
-    toText = toTextFromBoundedEnum SnakeLowerCase
-
-instance FromText ChangeChain where
-    fromText = fromTextToBoundedEnum SnakeLowerCase
-
--- | Purpose is a constant set to 44' (or 0x8000002C) following the BIP-44
--- recommendation. It indicates that the subtree of this node is used
--- according to this specification.
---
--- Hardened derivation is used at this level.
-purposeIndex :: Word32
-purposeIndex = 0x8000002C
-
--- | One master node (seed) can be used for unlimited number of independent
--- cryptocoins such as Bitcoin, Litecoin or Namecoin. However, sharing the
--- same space for various cryptocoins has some disadvantages.
---
--- This level creates a separate subtree for every cryptocoin, avoiding reusing
--- addresses across cryptocoins and improving privacy issues.
---
--- Coin type is a constant, set for each cryptocoin. For Cardano this constant
--- is set to 1815' (or 0x80000717). 1815 is the birthyear of our beloved Ada
--- Lovelace.
---
--- Hardened derivation is used at this level.
-coinTypeIndex :: Word32
-coinTypeIndex = 0x80000717
-
--- | Generate a new key from seed. Note that the @depth@ is left open so that
--- the caller gets to decide what type of key this is. This is mostly for
--- testing, in practice, seeds are used to represent root keys, and one should
--- use 'generateKeyFromSeed'.
-unsafeGenerateKeyFromSeed
-    :: (Passphrase "seed", Passphrase "generation")
-        -- ^ The actual seed and its recovery / generation passphrase
-    -> Passphrase "encryption"
-    -> Key depth XPrv
-unsafeGenerateKeyFromSeed (Passphrase seed, Passphrase gen) (Passphrase pwd) =
-    let
-        seed' = invariant
-            ("seed length : " <> show (BA.length seed) <> " in (Passphrase \"seed\") is not valid")
-            seed
-            (\s -> BA.length s >= 16 && BA.length s <= 255)
-    in Key $ generateNew seed' gen pwd
-
--- | Generate a root key from a corresponding seed
--- The seed should be at least 16 bytes
-generateKeyFromSeed
-    :: (Passphrase "seed", Passphrase "generation")
-        -- ^ The actual seed and its recovery / generation passphrase
-    -> Passphrase "encryption"
-    -> Key 'RootK XPrv
-generateKeyFromSeed = unsafeGenerateKeyFromSeed
-
--- | Derives account private key from the given root private key, using
--- derivation scheme 2 (see <https://github.com/input-output-hk/cardano-crypto/ cardano-crypto>
--- package for more details).
---
--- NOTE: The caller is expected to provide the corresponding passphrase (and to
--- have checked that the passphrase is valid). Providing a wrong passphrase will
--- not make the function fail but will instead, yield an incorrect new key that
--- doesn't belong to the wallet.
-deriveAccountPrivateKey
-    :: Passphrase "encryption"
-    -> Key 'RootK XPrv
-    -> Index 'Hardened 'AccountK
-    -> Key 'AccountK XPrv
-deriveAccountPrivateKey (Passphrase pwd) (Key rootXPrv) (Index accIx) =
-    let
-        purposeXPrv = -- lvl1 derivation; hardened derivation of purpose'
-            deriveXPrv DerivationScheme2 pwd rootXPrv purposeIndex
-        coinTypeXPrv = -- lvl2 derivation; hardened derivation of coin_type'
-            deriveXPrv DerivationScheme2 pwd purposeXPrv coinTypeIndex
-        acctXPrv = -- lvl3 derivation; hardened derivation of account' index
-            deriveXPrv DerivationScheme2 pwd coinTypeXPrv accIx
-    in
-        Key acctXPrv
-
--- | Derives address private key from the given account private key, using
--- derivation scheme 2 (see <https://github.com/input-output-hk/cardano-crypto/ cardano-crypto>
--- package for more details).
---
--- It is preferred to use 'deriveAddressPublicKey' whenever possible to avoid
--- having to manipulate passphrases and private keys.
---
--- NOTE: The caller is expected to provide the corresponding passphrase (and to
--- have checked that the passphrase is valid). Providing a wrong passphrase will
--- not make the function fail but will instead, yield an incorrect new key that
--- doesn't belong to the wallet.
-deriveAddressPrivateKey
-    :: Passphrase "encryption"
-    -> Key 'AccountK XPrv
-    -> ChangeChain
-    -> Index 'Soft 'AddressK
-    -> Key 'AddressK XPrv
-deriveAddressPrivateKey
-        (Passphrase pwd) (Key accXPrv) changeChain (Index addrIx) =
-    let
-        changeCode =
-            fromIntegral $ fromEnum changeChain
-        changeXPrv = -- lvl4 derivation; soft derivation of change chain
-            deriveXPrv DerivationScheme2 pwd accXPrv changeCode
-        addrXPrv = -- lvl5 derivation; soft derivation of address index
-            deriveXPrv DerivationScheme2 pwd changeXPrv addrIx
-    in
-        Key addrXPrv
-
--- | Derives address public key from the given account public key, using
--- derivation scheme 2 (see <https://github.com/input-output-hk/cardano-crypto/ cardano-crypto>
--- package for more details).
---
--- This is the preferred way of deriving new sequential address public keys.
-deriveAddressPublicKey
-    :: Key 'AccountK XPub
-    -> ChangeChain
-    -> Index 'Soft 'AddressK
-    -> Key 'AddressK XPub
-deriveAddressPublicKey (Key accXPub) changeChain (Index addrIx) =
-    fromMaybe errWrongIndex $ do
-        let changeCode = fromIntegral $ fromEnum changeChain
-        changeXPub <- -- lvl4 derivation in bip44 is derivation of change chain
-            deriveXPub DerivationScheme2 accXPub changeCode
-        addrXPub <- -- lvl5 derivation in bip44 is derivation of address chain
-            deriveXPub DerivationScheme2 changeXPub addrIx
-        return $ Key addrXPub
-  where
-    errWrongIndex = error $
-        "Cardano.Wallet.Primitive.AddressDerivation.deriveAddressPublicKey \
-        \failed: was given an hardened (or too big) index for soft path \
-        \derivation ( " ++ show addrIx ++ "). This is either a programmer \
-        \error, or, we may have reached the maximum number of addresses for \
-        \a given wallet."
 
 -- | Convert a public key to an 'Address' depending on a particular target.
 --
