+++ lib/core/test/integration/Test/Integration/Scenario/API/Addresses.hs
@@ -12,7 +12,7 @@ import Prelude
 
 import Cardano.Wallet.Api.Types
     ( ApiAddress, ApiTransaction, ApiWallet )
-import Cardano.Wallet.Primitive.AddressDiscovery
+import Cardano.Wallet.Primitive.AddressDiscovery.Sequential
     ( defaultAddressPoolGap, getAddressPoolGap )
 import Cardano.Wallet.Primitive.Types
     ( AddressState (..), DecodeAddress (..), EncodeAddress (..) )
