+++ lib/core/test/integration/Test/Integration/Scenario/CLI/Addresses.hs
@@ -11,7 +11,7 @@ import Prelude
 
 import Cardano.Wallet.Api.Types
     ( ApiAddress, getApiT )
-import Cardano.Wallet.Primitive.AddressDiscovery
+import Cardano.Wallet.Primitive.AddressDiscovery.Sequential
     ( defaultAddressPoolGap, getAddressPoolGap )
 import Cardano.Wallet.Primitive.Types
     ( AddressState (..), DecodeAddress (..), EncodeAddress (..) )
