+++ lib/core/test/unit/Cardano/Wallet/DB/StateMachine.hs
@@ -58,7 +58,7 @@ import Cardano.Wallet.DummyTarget.Primitive.Types
     ( DummyTarget, Tx (..) )
 import Cardano.Wallet.Primitive.AddressDerivation
     ( Key, XPrv, deserializeXPrv )
-import Cardano.Wallet.Primitive.AddressDiscovery
+import Cardano.Wallet.Primitive.AddressDiscovery.Sequential
     ( SeqState (..) )
 import Cardano.Wallet.Primitive.Model
     ( Wallet )
